Use the rdma_cm_id from CMINFO rather than from the event
in cm_close, the rdma_cm_id from dev->cm is used to close the connection
for both the server and the client. However, in cm_open_server the id
from the cm event is used. The pointer for the qp seems to change for
the event's id and causes a segfault in rdma_destroy_qp on the server
side.